refactor: exercises to follow new standard
