+++ src/lib/db/index.ts
@@ -1,10 +1,9 @@
-import { drizzle, type BetterSQLite3Database } from 'drizzle-orm/better-sqlite3'
+import { drizzle } from 'drizzle-orm/better-sqlite3'
 import Database from 'better-sqlite3'
 
 import * as AuthSchema from "./schema/auth"
 import * as MuscleSchema from "./schema/muscles"
 import * as ExerciseSchema from "./schema/exercises/schema"
-import * as ExerciseMusclesSchema from "./schema/exercises/muscleGroups"
 import * as WorkoutSchema from "./schema/workouts/schema"
 
 const sqlite = new Database(process.env.DB_URL!)
@@ -13,7 +12,6 @@ export const db = drizzle(sqlite, {
     ...AuthSchema,
     ...MuscleSchema,
     ...ExerciseSchema,
-    ...ExerciseMusclesSchema,
     ...WorkoutSchema
   }
 })
