+++ src/lib/db/schema/workouts/schema.ts
@@ -1,0 +1,4 @@
+export * from "./workouts.schema"
+export * from "./workoutsOnMuscles.schema"
+export * from "./workoutExercises.schema"
+export * from "./worktoutSets.schema"
