+++ src/lib/db/schema/workouts/workouts.schema.ts
@@ -31,7 +31,3 @@ export const workoutsRelations = relations(workouts, ({ one, many }) => ({
 
   workoutMuscles: many(workoutsOnMuscles)
 }))
-
-export * from "./workoutsOnMuscles.schema"
-export * from "./workoutExercises.schema"
-export * from "./worktoutSets.schema"
